requirements.txt
numpy>=1.26
matplotlib>=3.8
scikit-learn>=1.3
scipy>=1.11
ripser>=0.6
persim>=0.3
gudhi>=3.10
trimesh>=4.0
pillow>=10.0
pytest>=7.0
jupyter>=1.0
notebook>=7.0
nbformat>=5.0
